src/api/requirements.txt
Flask==2.2.2
Gunicorn
mysql-connector-python
osmnx
scikit-learn
